[#158] refactor : 피드 페이지 중복 제거를 위한 FeedPageTemplate 추출
